fixed ping, fixed error messages

+++ src/SystemHealth.py
@@ -171,8 +171,8 @@ def task_cpu(image_name):
     for i in range(2):
 
         for p in c.Win32_PerfRawData_PerfProc_Process(name=image_name):
-            id = long(p.IDProcess)
-            n1, d1 = long(p.PercentProcessorTime), long(p.Timestamp_Sys100NS)
+            id = int(p.IDProcess)
+            n1, d1 = int(p.PercentProcessorTime), int(p.Timestamp_Sys100NS)
             n0, d0, so_far = process_info.get(id, (0, 0, []))
 
             try:
